pvDatabaseRPC: Replace "move" by multiple operations
More development of pvDatabaseRPC example. In place of a single, simple
"move" operation it now has a complex state machine and set of services:
Configure, run (non-blocking), pause, resume, scan (blocking),
rewind, stop and abort.

A state field has been added to the PV. Each of the 3 main fields (the
setpoint, readback and state) has a principal "value" field and time
stamp, in addition to a top-level "timeStamp" field. A client supporting
the operations has been provided.

It illustrates synchronous and asynchronous RPC services, selecting
services based on the supplied pvRequest and using EPICS V4 to talk to
an existing object which has no knowledge of EPICS and call its methods,
so creating distributed objects.

Documentation has been updated and help/usage for the operations has
been added to the client.